Restructure for consistency between test src and src
